demo/requirements.txt
liac-arff==2.1.0
tabulate==0.7.5
scikit-learn==0.17.1
scikit-neuralnetwork==0.6.1
